serpent: Do not distinguish between 2D and 3D entities
It makes very little sense for us to distinguish between a 2D or
3D entity, as a basic Entity has no real logic or display code.
Thus they become the same storage type.

In time we may need to revisit this, but for now we'll start adding
some kind of tagging so that the renderers can discover the objects
they need to be looking out for.

Likely this will be implemented in a tag bucket within the scene.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>