input: Dispatch proper event types as structs
These wrap the SDL underlying types but make it significantly easier
to use the library, without having to do the switcheroo logic seen
with typical SDL patterns.

This abstracts away SDL enough that users of the framework don't need
to understand SDL or know it is being used, just know how basic
event handling works. Kinda JavaIsh, but not.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>