sprite: Hack to get texture rendering working
Again, super hacky. But this now has us rendering a sample texture for
now, which is miles ahead of what we could do before. We have a PosUVVertex
type now that contains UV coords as well as actual positional data.

Next step is to flip the normals to get the image showing the right way
around.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>