core: Add the new Group type
A Group provides a mechanism to chain like-systems or stages within a
single stage of execution. The Group may only have Processors of the
same (compiler-enforced) Data Policy, ensuring a promise of data integrity.

A Group may either execute sequentially, in the original insertation order,
or in parallel. Parallel system execution is only supported when using a
ReadOnly DataPolicy, and the `parallel` property has been set to `true`.

The Group may only be executed using a `TaskPool` provided from the running
context.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>