display: Standardize on OpenGL for now
It would appear that the OpenGL implementation in bgfx is completely solid,
however the Vulkan implementation requires too many corner cases. The plan
therefore, is to get feature-complete with OpenGL, and then re-attain feature
parity with Vulkan later on.

Interesting corner cases include vsync being lost across bgfx_reset calls,
or fullscreen causing a display lock. We'll get there. Note this is no fault
of bgfx, we need to rebase on newer versions and then ensure we correctly
take all the right paths.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>