graphics: Remove archaic Renderer foreach
We're trying to get all modern and sexy, and we can't exactly do that
with the old pipeline approach holding us back. Globally we're only
using bgfx with no additional abstractions so we don't really need
Renderer notions.

We now expost Pre/Render/Post groups, which will be extensible in future
and force read-only access as one would expect. This will enable us to
have multithreaded rendering in future with very little work. Yay us.

The Pre-Render group is currently just responsible for kicking off bgfx
with clearing the screen, etc. The Render group is where we'll add processors
so that they may get a view of the entity data and do something halfways
intelligent. The Post-Render group is where we flush the frame, but we
*may* also allow post-processing here at some point.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>