core: Introduce the new View type
Right now this is a heavy weight class so we *may* turn it into a
simpler struct. Anyway, the idea of the view is to abstract access
to the entity and component managers and indirect their package
level functions.

This will allow us to restrict all API usage to the View type,
including any modification or iteration of entities.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>