core: Introduce a new TransformComponent
As and when we fix up the camera and sprite APIs, the TransformComponent
will be expanded upon. For now, it will be used to describe the space
position of any entity within the game. This will allow us to start
fleshing out more APIs as we go along to get the camera system working
again.

Yay for ECS..?

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>