entity: Ensure all new Entity's have at least a TransformComponent
This takes a bit of inspiration from other engines that once used the
idea of GameObjects (and similarly named). We want our entities to be
pretty much instantly usable, and ensure that all our processors can
safely assume the presence of a TransformComponent.

Additionally, this means our Camera system will safely be able to
iterate all entities with TransformComponent and modify their renderable
positions.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>