serpent: Begin the new Tiled module
The serpent.tiled module (namespace subject to change) will form the
basis of our tilemap capabilities. Currently we have a very rudimentary
understanding of tilemaps, in that they sequentially contain layers of
information in uint32_t blobs.

Each layer's data is composed of a 1D uint32_t array which will in turn
have X,Y values corresponding to the GUID.

Ideally we could do with a sparse integer set per layer to optimise away
unused tiles, but this will do for now.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>