renderer: Ensure tiled renderer logic-loop is in place
For now we step through the layers in ascending order, but in future we'll
respect what the map actually says. Additionally we only know how to render
orthogonal maps right now.

Basic logic is to step the rows then the columns and extract the gid from
the layer data. Then draw the correct texture region at the right place.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>