batch: Introduce new TexturedQuad type
This will be used in internal storage to form a queue of drawing operations
sorted by texture. It's not being used just yet but we're trying to keep the
changes small and managable. Eventally we'll only pull drawing ops out of the
queue and disallow direct drawing, in order to optimise the number of draw
calls and texture switches.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>