ManyHacks: Start bringing back the Pipeline idea
We need to sort our entities before we begin to draw them, otherwise our
transparency is never going to work. Thus we need a single Pipeline that
will house renderers, which in turn will draw any incoming entities that
have the specified component. Notably we become renderable through the
registration of one chief component right now, so lets roll with it.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>