frame: Begin new simplistic framepacket notion
The FramePacket is a preallocated set of ringbuffers that allow us to
stash all renderable entities for the upcoming draw. If it's not in the
draw list, it won't be processed for rendering.

This will allow us to cull off-screen objects automatically, and hopefully
somewhat optimise our flow a bit.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>