graphics: Abstract out a new Renderer notion
These renderers will be responsible for querying the visibility of a given
thing according to the known dimensions. This becomes a component level
responsibility for implementing culling, by querying and submitting the
entities in the world for rendering.

Additionally this will allow us to implement chunked processing at the
renderer level for more complex renderers.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>