batch: Hax: Reuse the locking ringbuffer for now to simplify batching
We reuse most of the notions so lets just reuse the ringbuffer until such
point we convert QuadBatch into a smart dispatch service.

Note we still have an issue whereby we allocate all these transient index
buffers so we have to have more drawcalls to stop a segfault. We'll fix
this up as the pipeline evolves.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>