context: Add internal notion of deltaTime
This may be better off being in the frame packet in future but it does
now allow us to track the time passed for the current frame vs the last
frame.

Realistically we also need a ticks count that measures how far we are in
total so that we can provide tweening, but we'll get there. We'll need this
for dropped frames.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>