display: Be responsible for bootstrapping the pipeline
The pipeline can only be bootstrapped once the display is up
and running, so we ask the display to do this as soon as we
get a window. Next up is making the pipeline responsible for
all bgfx initialisation and taking that out of the display
code to unspaghetti it.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>