core: Introduce the new GreedyArray type
This is a self managed array that automatically grows in power-of-2
increments (when used correctly) to greedily reallocate the underlying
cache-coherent buffer. It can grow to a maximum size or an unbounded one.

There are many places we're manually doing some kind of array abuse in
the codebase so its about time we standardised it.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>