archetype: Introduce the new 'Archetype' super-entity-type
This provides grouping functionality for similar entities. An archetype
is composed of an exact set of components, and any entitiy that has
exactly the same components will join that archetype.

Upon joining, any relevant storage data from an older archetype will
be cloned, preserving it over component changes. The internal storage
is managed through a list of pages.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>