entitymanager: replace old EC with EntityManager
The EntityManager is responsible for the entire contents of a 'world',
providing automatic archetype management and component management.

Entities may be manipulated via several APIs but it is highly recommended
to use the View APIs instead.

The EntityManager is highly performant in terms of Read queries, and still
pretty darn good at creations. All similar entities are grouped into an
archetype, making sure multiple component queries are dirt cheap as the
entity record has the same index in the whole page.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>