dub: Statically link D runtime in release mode
Chances are we want a simple deployment of Serpent applications, so shipping
the runtime as a separate library makes that somewhat more difficult. In release
mode we'll rely on statically linked runtime to avoid being broken by upgrades
to the host D runtime.

In theory this raises the potential for security issues, however LDC ABI
breaks across releases, so we can't just hope to redirect the .so for
improved security, ala LSI style.

Developers will regularly need to rebuild their games with LDC should security
issues arise in the LDC distribution.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>