core: Add support for custom 'free' functions on components
This rather complex change is provided so that we can stash custom pointers
into our struct-storage and ensure it is correctly freed when needed. As
such this helps us with Chipmunk2D integration by ensuring we only free a
cpBody when we're really done with it, i.e. removing the component or killing
the entity.

As a final good grace, we ensure to clear the entitymanager, and any associated
data/memory, on context run completion.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>