entitymanager: Avoid catastrophic off-by-one errors
The signature may at times include a '0' field to indicate it is unset.
As such, we MUST NOT include these in the fetch for component manifests
otherwise we'll try to retrieve index position -1, which is completely
invalid.

Filter the signature into component manifests that *CAN* be valid, and
yay, we fix debug mode.

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>