graphics: Allow recolouring textures on the fly through shader
This is just a basic colour multiplication which will allow us to
do simple recolouring. Right now we don't expose the texture colour
anywhere but we'll need it in the near future to allow animating
opacity and such.

We'll also want to be able to recolour whole framebuffers to add
very simple scene-wide visual effects (dusk/dawn/etc)

Signed-off-by: Ikey Doherty <ikey.doherty@lispysnake.com>